refactor: アプリケーションの初期化処理を整理し、表示エリアとコピー機能を分離
