+++ js/displayAreaHandler.js
@@ -1,0 +1,20 @@
+// displayAreaHandler.js
+// 表示エリアの初期化・再描画・ドラッグ＆ドロップハンドラ登録を担当
+
+import { renderLines } from './renderLines.js';
+import { moveLine } from './linesData.js';
+import { attachDragDropHandlers } from './dragDropHandler.js';
+
+/**
+ * 表示エリアの初期化と再描画、D&Dハンドラ登録を行う
+ */
+export function initDisplayArea() {
+    const displayArea = document.getElementById('display-area');
+    if (displayArea) {
+        const rerender = () => {
+            renderLines(displayArea);
+            attachDragDropHandlers(displayArea, moveLine, rerender);
+        };
+        rerender();
+    }
+}
